// playground/src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type ApplicationQueries = {
  __typename?: 'ApplicationQueries';
  /** Find an articles content by id. */
  findArticlesContent?: Maybe<Articles>;
  /** Find an asset by id. */
  findAsset?: Maybe<Asset>;
  /** Find an The relationship of Hotels and introductions content by id. */
  findColumnsContent?: Maybe<Columns>;
  /** Find an Hotels content by id. */
  findHotelsContent?: Maybe<Hotels>;
  /** Find an introductions content by id. */
  findIntroductionsContent?: Maybe<Introductions>;
  /** Find an Pages content by id. */
  findPagesContent?: Maybe<Pages>;
  /** Find an Posts content by id. */
  findPostsContent?: Maybe<Posts>;
  /** Find an test content by id. */
  findTestContent?: Maybe<Test>;
  /** Query articles content items. */
  queryArticlesContents?: Maybe<Array<Articles>>;
  /** Query articles content items with total count. */
  queryArticlesContentsWithTotal?: Maybe<ArticlesResultDto>;
  /** Get assets. */
  queryAssets: Array<Asset>;
  /** Get assets and total count. */
  queryAssetsWithTotal: AssetResultDto;
  /** Query The relationship of Hotels and introductions content items. */
  queryColumnsContents?: Maybe<Array<Columns>>;
  /** Query The relationship of Hotels and introductions content items with total count. */
  queryColumnsContentsWithTotal?: Maybe<ColumnsResultDto>;
  /** Query content items by IDs across schemeas. */
  queryContentsByIds: Array<AllContents>;
  /** Query Hotels content items. */
  queryHotelsContents?: Maybe<Array<Hotels>>;
  /** Query Hotels content items with total count. */
  queryHotelsContentsWithTotal?: Maybe<HotelsResultDto>;
  /** Query introductions content items. */
  queryIntroductionsContents?: Maybe<Array<Introductions>>;
  /** Query introductions content items with total count. */
  queryIntroductionsContentsWithTotal?: Maybe<IntroductionsResultDto>;
  /** Query Pages content items. */
  queryPagesContents?: Maybe<Array<Pages>>;
  /** Query Pages content items with total count. */
  queryPagesContentsWithTotal?: Maybe<PagesResultDto>;
  /** Query Posts content items. */
  queryPostsContents?: Maybe<Array<Posts>>;
  /** Query Posts content items with total count. */
  queryPostsContentsWithTotal?: Maybe<PostsResultDto>;
  /** Query test content items. */
  queryTestContents?: Maybe<Array<Test>>;
  /** Query test content items with total count. */
  queryTestContentsWithTotal?: Maybe<TestResultDto>;
};
// ...
export type Articles = Content & {
  __typename?: 'Articles';
  /** The timestamp when the object was created. */
  created: Scalars['Instant']['output'];
  /** The user who created the object. */
  createdBy: Scalars['String']['output'];
  /** The user who created the object. */
  createdByUser: User;
  /** The data of the content. */
  data: ArticlesDataDto;
  /** The data of the content. */
  data__dynamic?: Maybe<Scalars['JsonScalar']['output']>;
  /** The edit token. */
  editToken?: Maybe<Scalars['String']['output']>;
  /** The flat data of the content. */
  flatData: ArticlesFlatDataDto;
  /** The ID of the object (usually GUID). */
  id: Scalars['String']['output'];
  /** The timestamp when the object was updated the last time. */
  lastModified: Scalars['Instant']['output'];
  /** The user who updated the object the last time. */
  lastModifiedBy: Scalars['String']['output'];
  /** The user who updated the object the last time. */
  lastModifiedByUser: User;
  /** The new status of the content. */
  newStatus?: Maybe<Scalars['String']['output']>;
  /** The status color of the content. */
  newStatusColor?: Maybe<Scalars['String']['output']>;
  /** The status of the content. */
  status: Scalars['String']['output'];
  /** The status color of the content. */
  statusColor: Scalars['String']['output'];
  /** The URL to the content. */
  url: Scalars['String']['output'];
  /** The version of the objec. */
  version: Scalars['Int']['output'];
};
// ...
export type ArticlesFlatDataDto = {
  __typename?: 'ArticlesFlatDataDto';
  content?: Maybe<Scalars['String']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  slug?: Maybe<Scalars['String']['output']>;
};

/** List of articles items and total count. */
export type ArticlesResultDto = {
  __typename?: 'ArticlesResultDto';
  /** The contents. */
  items?: Maybe<Array<Articles>>;
  /** The total count of  contents. */
  total: Scalars['Int']['output'];
};

/** An asset */
export type Asset = {
  __typename?: 'Asset';
  /** The timestamp when the object was created. */
  created: Scalars['Instant']['output'];
  /** The user who created the object. */
  createdBy: Scalars['String']['output'];
  /** The user who created the object. */
  createdByUser: User;
  /** The edit token. */
  editToken?: Maybe<Scalars['String']['output']>;
  /** The hash of the file. Can be null for old files. */
  fileHash: Scalars['String']['output'];
  /** The file name of the asset. */
  fileName: Scalars['String']['output'];
  /** The size of the file in bytes. */
  fileSize: Scalars['Int']['output'];
  /** The file type (file extension) of the asset. */
  fileType: Scalars['String']['output'];
  /** The version of the file. */
  fileVersion: Scalars['Int']['output'];
  /** The ID of the object (usually GUID). */
  id: Scalars['String']['output'];
  /**
   * Determines if the uploaded file is an image.
   * @deprecated Use 'type' field instead.
   */
  isImage: Scalars['Boolean']['output'];
  /** True, when the asset is not public. */
  isProtected: Scalars['Boolean']['output'];
  /** The timestamp when the object was updated the last time. */
  lastModified: Scalars['Instant']['output'];
  /** The user who updated the object the last time. */
  lastModifiedBy: Scalars['String']['output'];
  /** The user who updated the object the last time. */
  lastModifiedByUser: User;
  /** The asset metadata. */
  metadata?: Maybe<Scalars['JsonScalar']['output']>;
  /** The type of the image. */
  metadataText: Scalars['String']['output'];
  /** The mime type. */
  mimeType: Scalars['String']['output'];
  /** The ID of the parent folder. Empty for files without parent. */
  parentId: Scalars['String']['output'];
  /**
   * The height of the image in pixels if the asset is an image.
   * @deprecated Use 'metadata' field instead.
   */
  pixelHeight?: Maybe<Scalars['Int']['output']>;
  /**
   * The width of the image in pixels if the asset is an image.
   * @deprecated Use 'metadata' field instead.
   */
  pixelWidth?: Maybe<Scalars['Int']['output']>;
  /** The file name as slug. */
  slug: Scalars['String']['output'];
  /** The source URL of the asset. */
  sourceUrl?: Maybe<Scalars['String']['output']>;
  /** The asset tags. */
  tags: Array<Scalars['String']['output']>;
  /** The thumbnail URL to the asset. */
  thumbnailUrl?: Maybe<Scalars['String']['output']>;
  /** The type of the asset. */
  type: AssetType;
  /** The URL to the asset. */
  url: Scalars['String']['output'];
  /** The version of the objec. */
  version: Scalars['Int']['output'];
};
// ...
export type ChaptersComponent = Component & {
  __typename?: 'ChaptersComponent';
  articles?: Maybe<Array<Articles>>;
  /** The ID of the schema. */
  schemaId: Scalars['String']['output'];
  /** The name of the schema. */
  schemaName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type Hotels = Content & {
  __typename?: 'Hotels';
  /** The timestamp when the object was created. */
  created: Scalars['Instant']['output'];
  /** The user who created the object. */
  createdBy: Scalars['String']['output'];
  /** The user who created the object. */
  createdByUser: User;
  /** The data of the content. */
  data: HotelsDataDto;
  /** The data of the content. */
  data__dynamic?: Maybe<Scalars['JsonScalar']['output']>;
  /** The edit token. */
  editToken?: Maybe<Scalars['String']['output']>;
  /** The flat data of the content. */
  flatData: HotelsFlatDataDto;
  /** The ID of the object (usually GUID). */
  id: Scalars['String']['output'];
  /** The timestamp when the object was updated the last time. */
  lastModified: Scalars['Instant']['output'];
  /** The user who updated the object the last time. */
  lastModifiedBy: Scalars['String']['output'];
  /** The user who updated the object the last time. */
  lastModifiedByUser: User;
  /** The new status of the content. */
  newStatus?: Maybe<Scalars['String']['output']>;
  /** The status color of the content. */
  newStatusColor?: Maybe<Scalars['String']['output']>;
  /** Query The relationship of Hotels and introductions content items. */
  referencingColumnsContents?: Maybe<Array<Columns>>;
  /** Query The relationship of Hotels and introductions content items with total count. */
  referencingColumnsContentsWithTotal?: Maybe<ColumnsResultDto>;
  /** The status of the content. */
  status: Scalars['String']['output'];
  /** The status color of the content. */
  statusColor: Scalars['String']['output'];
  /** The URL to the content. */
  url: Scalars['String']['output'];
  /** The version of the objec. */
  version: Scalars['Int']['output'];
};
// ...
export type Introductions = Content & {
  __typename?: 'Introductions';
  /** The timestamp when the object was created. */
  created: Scalars['Instant']['output'];
  /** The user who created the object. */
  createdBy: Scalars['String']['output'];
  /** The user who created the object. */
  createdByUser: User;
  /** The data of the content. */
  data: IntroductionsDataDto;
  /** The data of the content. */
  data__dynamic?: Maybe<Scalars['JsonScalar']['output']>;
  /** The edit token. */
  editToken?: Maybe<Scalars['String']['output']>;
  /** The flat data of the content. */
  flatData: IntroductionsFlatDataDto;
  /** The ID of the object (usually GUID). */
  id: Scalars['String']['output'];
  /** The timestamp when the object was updated the last time. */
  lastModified: Scalars['Instant']['output'];
  /** The user who updated the object the last time. */
  lastModifiedBy: Scalars['String']['output'];
  /** The user who updated the object the last time. */
  lastModifiedByUser: User;
  /** The new status of the content. */
  newStatus?: Maybe<Scalars['String']['output']>;
  /** The status color of the content. */
  newStatusColor?: Maybe<Scalars['String']['output']>;
  /** Query The relationship of Hotels and introductions content items. */
  referencingColumnsContents?: Maybe<Array<Columns>>;
  /** Query The relationship of Hotels and introductions content items with total count. */
  referencingColumnsContentsWithTotal?: Maybe<ColumnsResultDto>;
  /** The status of the content. */
  status: Scalars['String']['output'];
  /** The status color of the content. */
  statusColor: Scalars['String']['output'];
  /** The URL to the content. */
  url: Scalars['String']['output'];
  /** The version of the objec. */
  version: Scalars['Int']['output'];
};
// ...
export type IntroductionsFlatDataDto = {
  __typename?: 'IntroductionsFlatDataDto';
  chapters?: Maybe<Array<ChaptersComponent>>;
  chapters__Dynamic?: Maybe<Scalars['JsonScalar']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  slug?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type IntroQuery = { __typename?: 'ApplicationQueries', intros?: Array<{ __typename?: 'Introductions', flatData: { __typename?: 'IntroductionsFlatDataDto', title?: string | null, description?: string | null, chapters?: Array<{ __typename?: 'ChaptersComponent', title?: string | null, articles?: Array<{ __typename?: 'Articles', id: string, flatData: { __typename?: 'ArticlesFlatDataDto', name?: string | null } }> | null }> | null } }> | null };

export type ArticleQueryVariables = Exact<{
  filter: Scalars['String']['input'];
}>;


export type ArticleQuery = { __typename?: 'ApplicationQueries', articles?: Array<{ __typename?: 'Articles', flatData: { __typename?: 'ArticlesFlatDataDto', name?: string | null, content?: string | null } }> | null };

export type SidebarQueryVariables = Exact<{
  filter: Scalars['String']['input'];
}>;


export type SidebarQuery = { __typename?: 'ApplicationQueries', sidebars?: Array<{ __typename?: 'Introductions', flatData: { __typename?: 'IntroductionsFlatDataDto', chapters?: Array<{ __typename?: 'ChaptersComponent', title?: string | null, articles?: Array<{ __typename?: 'Articles', id: string, flatData: { __typename?: 'ArticlesFlatDataDto', name?: string | null } }> | null }> | null } }> | null };